pip-requirements.txt
python-Levenshtein
Django==1.8.1
psycopg2==2.5
django-model-utils
dateutils
supervisor
ordered-set
rq>0.3.1
rq_scheduler>=0.5.2
django-rq
django-rq-dashboard
channels
asgi_redis
django-cacheops
django-passwords
django-timezone-field


